refactor: main.rs(4054行)を17モジュールに分割 (#30)
main.rsが肥大化していたため、機能ごとにモジュール分割:
- docker.rs: Docker接続・認証・イメージpull
- utils.rs: ステージ名決定・変数展開・共通ユーティリティ
- self_update.rs: セルフアップデート
- play.rs: Playbook実行
- build.rs: Dockerイメージビルド・buildx
- commands/: 各コマンドハンドラー(up,down,ps,logs,restart,stop,start,deploy,validate,exec,stage)

main.rsはCLI定義+ディスパッチのみ(~400行)に削減。
外部APIの変更なし、全テスト通過。

Closes #30
