+++ crates/fleetflow/src/commands/mod.rs
@@ -1,0 +1,11 @@
+pub mod deploy;
+pub mod down;
+pub mod exec;
+pub mod logs;
+pub mod ps;
+pub mod restart;
+pub mod stage;
+pub mod start;
+pub mod stop;
+pub mod up;
+pub mod validate;
